Updated default options
I ':sort'-ed the default options, and moved the initialization of
$nodename into it.

I also added 'Net::DNS' as a module, as there is some dns querying that
needs to be done if a KDC/LDAP server isn't specified on command line. I
may eventually pull the code for this into it's own file, but I haven't
decided yet.