Added discover_and_set_domain()
This function will discover the domain for the computer and set
'$domain' to it. I also added more code to the block that checks for the
domain and nodename, so it will call 'discover_and_set_domain' if the
domain isn't specified on the command line. Specifically, if, after
parsing the options, $#ARGV < 0 (nothing left on the command line), then
'discover_and_set_domain' is called.