Changed 'getSRVs' to 'get_SRVs', and other changes
Changed the name of 'getSRVs' to 'get_SRVs', for better readability and
consistency with the style of the rest of the file (consistency of style
is something I should work on...).

I also refactored 'get_SRVs' to use the new 'dns_search' function.

'get_SRVs' will return an array of hashes, instead of an array of
Net::DNS::RR objects, which is what 'dns_search' returns. The upside to
this is that the caller of 'get_SRVs' doesn't have to check the return
type of anything in the array, like they would with 'dns_search'. The
downside is that this is...inconsistent? This is where I kind of wish I
had someone else working on this with me. It's too easy to be
inconsistent if you don't have someone else complaining about your dumb
broken code.