Added or updated 'get_KDCs' and 'get_DCs'
I have added the 'get_KDCs' and 'get_DCs' functions. I have also
included code in the main scope of the function that handles the return
calls from those two functions. Also, I've translated some of the code
that was at the end of the 'getKDC' and 'getDC' functions in adjoin.sh
into code in the main scope.