Added 'construct_krb5_conf'
This function will construct the base level of what needs to be in
krb5.conf, put it in a temp file, and return the pathname to the temp
file, to be handled later as desired.

I also added code in the main scope that calls this function and sets
'krb5conf' to its return value.