Added KRB initializing code and 'find_forest'
I've added code that makes temporary file names for the kerberos admin
ticket and keytab, initializes the admin ticket, and finds the forest
name for the domain.

For this, I made the 'find_forest' function. It's not quite complete, as
it doesn't handle what happens when you don't pass it anything, but the
primary functionality is there.