Clarified function and variable names
I have changed some of the function names from 'get_FOO' to
'enumerate_FOO', because that's what they're doing, generating a list
(sort of) of all of the FOO that we have. However, 'get_SRVs' stays the
same, because it's actually retrieving SRV RRs from DNS, thus it is
"get"ting them, instead of enumerating them.