Added the work involved to find the site name
I have added in all of the functions needed to find the site name,
though I have no idea what that's supposed to be, just exactly. I have
put a big warning in the comment section of 'find_site', as I can't
really test its output.

I have also included the necessary code in the main scope that handles
things after a site name is found (again, can't test this, so no idea if
it works). I have placed a "TODO" along with this code, as it's a pretty
egregious violation of DRY, and looks ugly as sin.