Tried to tame some wild growth
I have added more of the functionality from adjoin.sh. Because I
couldn't stomach the thought of adding heaps more code to the main
scope, I made a new function called "sleuth_machine_bad_times" to try to
encapsulate it. While I tried to clean up the logic in the code, and
make it manageable, there's only so much I could do on the first pass,
so I tried to make this function as ugly as possible. There are warnings
in the comment header, the Input/Output sections of the comment header are
garbage, the name of the function is completely unhelpful, it takes a
somewhat ridiculous number of variables. I mean, this thing should
SCREAM "Refactor me!". When you see it, it's image should burn itself
into your eyes, it's message printed eternally in your brain. This
should invoke such a dark rage that you can't help but refactor it.

What I'm trying to say is that I did my best to try and make this port
an improvement on it's first pass, but there are some terrible blemishes
still, and they shouldn't be overlooked simply because they seem to be
functioning as expected (what is expected anyway...).

Really, I should do a rewrite once I have finished the translation and
understand what, just exactly, is trying to be accomplished. Add that to
the TODO list...