Added code to create and add the ldap machine account
I have added the code necessary to create and add an ldap account,
including the call to the function in the main scope. I have also
changed the '$object' variable to '$object_file', as that variable
contains the name of the object file to write to, not the contents of
the object itself.