Added note about kclient script in solaris in README
Searching around, I've found reference to a script called kclient from
Sun/Oracle. The script I have access to on a solaris machine called
kclient is similar to the one that I found reference to, except that it
doesn't seem to mention AD at all. So, theoretically, there's a script
called kclient out there that already handles this process. I don't know
exactly where that is, but it might be worth looking into for someone
that's living only in solaris (or maybe if that script also works on
other systems, that could be neat).