Added rest of main scope code
I have added the rest of the main scope code. This includes finding the
fqdn for the host, finalizing the machine account in ldap, writing out
the keytab file and setting up the kerberos configuration and keytab
file. Also, if on solaris, we'll call 'doIDMAPconfig'. Maybe that should
go in 'finalize_machine_account'...