Started minor refactoring
I have created "lib/DNS.pm" as the beginning work to modularize the
code. This commit doesn't contain any alteration to the subroutines
themselves, just a relocation.

I chose which routines to move based on how much work they did with DNS,
and, somewhat, what the routine was looking to accomplish. It's not a
perfect catagorization, necessarily, but it's a first step.