Removed almost all uses of '$krb5ccname'
I have discovered that setting $KRB5CCNAME in the environment via the
%ENV hash will be picked up throughout the whole program and by every
child (read: `system` and `qx` calls) process, unless otherwise dumped
by that process.

The downside to this process is that you can no longer just pass the
desired credentials cache path to each function that uses it. I can see
this maybe causing issues at some point, but it's extremelly trivial to
add support when it's needed, so I think this is ultimately a win for
maintainability.