Fixed modification ldap command in 'lib/Encryption.pm'
The LDIF object specification in 'deduce_and_set_enc_types' assumed that
there wasn't an 'msDS-SupportedEncryptionTypes' value in the machine
object, but from what I've read and seen, it's a part of a default
object. Thus, trying to add the value causes an error, and then the
script thinks that AES isn't supported (which it may or may not be).