Cleaned up code
I removed 'correct_idmap()', as I have no use for the code and no way of
testing it. It doesn't make any sense to keep code I neither need nor
use, so out it went.

I also cleaned up 'sleuth_machine_bad_times()', including renaming it to
'handle_preexisting_object()'. The machine accounts in my network don't
have sub-objects associated with '$netbios_nodename' (nor '$nodename' in
general), so there's no way to test this functionality, and furthermore
I don't need it. Until it's needed by someone, it shouldn't be included.

I renamed the '$leave' variable to '$part', because it made more sense
at the time. It is supposed to represent whether we are leaving the
domain or not. Thinking about it now, I will probably rename '$part' to
'$leave_domain', since that is explicit about what it represents. That
will be introduced in the next commit.