Cleaned up 'lib/Encryption.pm'
I removed all code that dealt with DES and ARCFOUR encryption types, as
those encryption types are not valid with the server that I'm working
with. In my continued effort to trim out any code that is either not
necessary or is untestable, this code got the axe.

In the 'deduce_linux_enc_info' and 'deduce_solaris_enc_info'
subroutines, I also pulled the code that tests the OS value into tiny
if-blocks at the top of the sub-routine. These if-blocks will return
from the routines if activated.

I also clarified various comments surrounding the affected code.