Clarified comments and code in plaidjoin.pl
I pulled the call to 'deduce_and_set_enc_types' out of
'finalize_machine_account', leaving only one call to the former in the
main scope of 'plaidjoin.pl'. I have placed the call to the former
before the call to the latter, so any desired or necessary effect is
still in place.

I have clarified a couple of comments in 'plaidjoin.pl'.