Restored 'kt_write' that utilizes 'Expect' module
Since my foray into using Authen::Krb5 to write out the keytab file for
the machine account has currently hit a dead end, I have restored the
parts of 'plaidjoin.pl' and 'lib/Kerberos.pm' that interact with
'kt_write'. That is, I have restored the necessary parts of
the two files so that the 'kt_write' sub-routine utilizes the 'Expect'
module.